clean up code & take into account errors

+++ node_cose_mbed2/include/obj-sec.h
@@ -22,7 +22,7 @@ void objsec_set_key(uint8_t * k);
  * @param szKey     Length of the key
  * @param prefsz    
  */
-size_t encrypt(uint8_t *buffer, uint16_t prefsz, const uint8_t *message, size_t len);
+int16_t encrypt(uint8_t *buffer, uint16_t prefsz, const uint8_t *message, size_t len);
 
 #ifdef __cplusplus
 }
