+++ node_cose_mbed2/src/obj-sec.c
@@ -39,7 +39,7 @@ void objsec_set_key(uint8_t *k){
     memcpy(key, k, OBJ_SEC_KEYSIZE);
 }
 
-size_t encrypt(uint8_t *buffer, uint16_t bufferSz, const uint8_t *message, size_t len) {
+int16_t encrypt(uint8_t *buffer, uint16_t bufferSz, const uint8_t *message, size_t len) {
   //HCOSE_ENCRYPT  COSE_Encrypt_Init(COSE_INIT_FLAGS flags, CBOR_CONTEXT_COMMA cose_errback * perr);
   //bool COSE_Encrypt_SetContent(HCOSE_ENCRYPT cose, const byte * rgbContent, size_t cbContent, cose_errback * errp);
   //bool COSE_Encrypt_encrypt(HCOSE_ENCRYPT cose, const byte * pbKey, size_t cbKey, cose_errback * perror);
